// vaccineApp/src/infrastructure/interfaces/create-edit-dosis-response.ts
// Generated by https://quicktype.io

export interface DosisByIDResponse {
    _id:             ID;
    name:            string;
    age_frequency:   string;
    expires_in_days: number;
    status:          boolean;
    rowReporte:      string;
    columReporte:    string;
    vaccine:         Vaccine;
}

export interface ID {
    $oid: string;
}

export interface Vaccine {
    _id:              ID;
    name:             string;
    description:      string;
    disease_prevents: string;
    application_age:  string;
    isChildren:       boolean;
    status:           boolean;
    dosis_ids:        any[];
}

 // Generated by https://quicktype.io

export interface DosisUpdateResponse {
    _id?:             ID;
    vacinne_id?:         string;
    error:      boolean;
    resp:       boolean;
    TypeError:  string;
    statusCode: string;
    ValueError: string;
    message:    string;
}

// Generated by https://quicktype.io

export interface DosisCreateResponse {
    id:              string;
    name:            string;
    vacinne_id:      string;
    age_frequency:   string;
    expires_in_days: number;
    status:          boolean;
}